+++ tests/testthat/test_Spectra.R
@@ -1,5 +1,5 @@
 test_that("query,Spectra works", {
-    expect_error(query(sps_dda), "single character")
+    expect_equal(query(sps_dda), sps_dda)
     expect_error(query(sps_dda, c("a", "b")), "single character")
     res <- query(sps_dda, "QUERY MS2DATA")
     expect_true(all(msLevel(res) == 2))
@@ -18,9 +18,10 @@ test_that(".query_spectra works", {
     expect_true(all(rtime(res) > 20))
     expect_true(all(acquisitionNum(res) >= 9 & acquisitionNum(res) <= 400))
 
-
-    res <- expect_error(.query_spectra(sps_dda, "QUERY * WHERE MS2PREC"), 
-                        "Non-numeric")
+    expect_error(.query_spectra(sps_dda, "QUERY * WHERE MS2PREC"),
+                 "non-numeric value")
+    expect_error(.query_spectra(sps_dda, "QUERY * WHERE MS2PREC = a"),
+                 "non-numeric value")
 
     ex_mz <- 304.1131
     mzr <- ex_mz + c(-1, 1) * ppm(ex_mz, 20)
@@ -33,27 +34,32 @@ test_that(".query_spectra works", {
     res <- .query_spectra(
         sps_dda, "QUERY * WHERE RTMIN=420 AND MS2PREC=304.1131:TOLERANCEPPM=20")
     expect_equal(length(res), 3)
-    
+
     res <- .query_spectra(sps_dda, "QUERY * WHERE CHARGE = 0")
     expect_true(all(precursorCharge(res) == 0))
-    
+
     res <- .query_spectra(sps_dda, "QUERY * WHERE CHARGE = NA OR -1")
     expect_true(all(is.na(precursorCharge(res)) | precursorCharge(res) == -1))
-    
+
     res <- .query_spectra(sps_dda, "QUERY * WHERE POLARITY = Positive")
-    expect_true(all(polarity(res) == 1)) 
-    
+    expect_true(all(polarity(res) == 1))
+
     res <- .query_spectra(
         sps_dda, "QUERY * WHERE MS2PROD=(100 OR 104):TOLERANCEPPM=5")
-    expect_true(containsMz(res, mz = 100, ppm = 5) ||
-                    containsMz(res, mz = 104, ppm = 5))
-    
+    expect_true(all(containsMz(res, mz = 100, ppm = 5) |
+                    containsMz(res, mz = 104, ppm = 5)))
+
+    res <- .query_spectra(
+        sps_dda, "QUERY * WHERE ms2prec = (99 or 90 or 89):tolerancemz=0.6")
+    expect_true(all(precursorMz(res) > 89))
+    expect_true(all(precursorMz(res) < 100))
+
     res <- .query_spectra(
         sps_dda, "QUERY * WHERE MS2NL=100:TOLERANCEPPM=5")
-    #expect_true(containsNeutralLoss(res, neutralLoss = 100, ppm = 5)) 
-    #res ha 0 spectra and containsNeutralLoss(res, neutralLoss = 100, ppm = 5)
-    # throws the error : Errore in value[[1L]] : subscript fuori limite. Is this 
-    # ok or should containsNeutralLoss return a logical value?
+##     #expect_true(containsNeutralLoss(res, neutralLoss = 100, ppm = 5))
+##     #res ha 0 spectra and containsNeutralLoss(res, neutralLoss = 100, ppm = 5)
+##     # throws the error : Errore in value[[1L]] : subscript fuori limite. Is this
+##     # ok or should containsNeutralLoss return a logical value?
 
 })