+++ NEWS.md
@@ -1,5 +1,11 @@
 # SpectraQL 0.98
 
+## Changes in 0.98.2
+
+- Change from `filterPrecursorMz` to `filterPrecursorMzValues`.
+- Support `"OR"` in `"MS2PREC"`, `MS1MZ`, `MS2PROD`.
+- Add `MS1MZ`.
+
 ## Changes in 0.98.1
 
 - Add functionality to extract the requested data from a `Spectra`.